+++ js/history.js
@@ -1,0 +1,3 @@
+import { timeline } from './modules/timeline.js';
+
+timeline();
