--- js/main.js
@@ -1,103 +1,0 @@
-//Plyr - Video Player
-(() => {
-    const videos = document.querySelectorAll('video');
-    
-    videos.forEach(video => {
-        const player = new Plyr(video, {
-            settings: [
-                'play-large',
-                'play',    
-                'progress', 
-                'current-time',
-                'mute',     
-                'volume',
-                'fullscreen' 
-            ]
-        }); //Sets controls to remove defaults to get rid of settings
-    });
-})();
-
-
-//----------------------------------------------------------------------------------//
-
-
-//Timeline
-(() => {
-    gsap.registerPlugin(ScrollTrigger);
-    
-    const timeline = document.querySelector('#timeline');
-    const progress = document.querySelector('#timeline-progress');
-    const timelineItems = document.querySelectorAll('.timeline-item');
-
-    //Progress bar
-    gsap.to(progress, {
-        height: '100%',
-        ease: 'none',
-        scrollTrigger: {
-            trigger: timeline,
-            start: 'top center',
-            end: 'bottom center',
-            scrub: true
-        }
-    });
-
-    //Dot triggers
-    timelineItems.forEach((item) => {
-        const dot = item.querySelector('.timeline-dot');
-        const content = item.querySelectorAll('.timeline-content div');
-
-        ScrollTrigger.create({
-            trigger: dot,
-            start: 'center center',
-            onEnter: () => {
-                gsap.to(content, {
-                    opacity: 1,
-                    x: 0,
-                    duration: 0.1,
-                    stagger: 0.1
-                });
-                item.classList.add('active');
-            },
-            onLeaveBack: () => {
-                gsap.to(content, {
-                    opacity: 0,
-                    x: 0,
-                    duration: 0.1,
-                    stagger: 0.1
-                });
-                item.classList.remove('active');
-            }
-        });
-    });
-})();
-
-const questions = document.querySelectorAll('.faq-question');
-
-questions.forEach(question => {
-    question.addEventListener('click', () => {
-        const card = question.closest('.faq-card');
-        
-        document.querySelectorAll('.faq-card.active').forEach(activeCard => {
-            if (activeCard !== card) activeCard.classList.remove('active');
-        });
-        
-        card.classList.toggle('active');
-    });
-});
-
-const missionButtons = document.querySelectorAll('.read-more-btn, .arrow');
-
-missionButtons.forEach(element => {
-    element.addEventListener('click', () => {
-        const missionCard = element.closest('.mission');
-        const button = missionCard.querySelector('.read-more-btn');
-        
-        missionCard.classList.toggle('expanded');
-        
-        if (missionCard.classList.contains('expanded')) {
-            button.textContent = 'Read Less';
-        } else {
-            button.textContent = 'Read More';
-        }
-    });
-});
